backend/requirements.txt
# Flask et extensions principales
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-Caching==2.1.0

# Base de données et drivers
PyMySQL==1.1.0
cryptography==41.0.7

# Validation et sérialisation
marshmallow==3.20.1

# Utilitaires et configuration
python-dotenv==1.0.0
Werkzeug==3.0.1
click==8.1.7

# Cache et Redis
redis==5.0.1

# Email et sécurité
itsdangerous==2.1.2
bcrypt==4.1.2

# Logging et monitoring
python-json-logger==2.0.7

# Tests
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0

# Production
gunicorn==21.2.0 